progress upto hyphenation
