+++ rtx_package/src/engine.rs
@@ -12,7 +12,7 @@ mod tex_glue;
 mod tex_hyphenation;
 mod tex_file_io;
 mod tex_inserts;
-// mod tex_job;
+mod tex_job;
 mod tex_kern;
 // mod tex_logic;
 // mod tex_macro;
@@ -48,7 +48,6 @@ mod tex_math_accents;
 mod tex_math_mode;
 mod tex_math_fork;
 mod tex_math_style;
-mod tex_misc_tweaks;
 mod tex_para;
 mod tex_registers;
 mod tex_scripts;
