+++ rtx_package/src/engine/plain.rs
@@ -23,4 +23,6 @@ LoadDefinitions!({
       } else { Ok(SymHashMap::default()) }
     });
 
+
+  Let!("\\empty", "\\lx@empty");
 });
