+++ rtx_package/src/engine/tex_special_chars.rs
@@ -76,8 +76,6 @@ LoadDefinitions!({
     document.get_node_mut().add_child(&mut replacement)?;
   });
 
-  DefMacro!("\\hiderel{}", "#1");    // Just ignore, for now...
-
   // RIGHTWARDS ARROW??? a bit more explicitly relation-like?
   DefMath!("\\to", None, "\u{2192}", role => "ARROW");
